Upgrades fonctionnels, jeu opérationnel et fin prêt !

+++ Assets/Scripts/Upgrades/Projectile.cs
@@ -3,14 +3,14 @@ using UnityEngine;
 
 public class Projectile : MonoBehaviour {
 
-    public int Degat = 2;
+    public float Degat = 2;
 
     private void OnTriggerEnter(Collider other) {
         if (other.CompareTag("Ennemy")) { // Ennemy touché
             Debug.Log("ENNEMY TOUCHE !!!!!!!!!!!!!!!!");
 
             // On applique les dégats à l'ennemy
-            other.GetComponent<Ennemy>().Stats.Health -= Degat;
+            other.GetComponent<Ennemy>().Stats.Health -= (int)Degat;
             Destroy(gameObject);
 
         }
